+++ packages/eyes-sdk-core/test/unit/fluent/CheckSettingsUtils.spec.js
@@ -431,6 +431,17 @@ describe('CheckSettingsUtils', () => {
     assert.deepStrictEqual(checkWindowConfiguration.variationGroupId, 'variant-id')
   })
 
+  it('toCheckWindowConfiguration handles pageId', () => {
+    const checkSettings = {pageId: 'my-page'}
+
+    const checkWindowConfiguration = CheckSettingsUtils.toCheckWindowConfiguration({
+      checkSettings,
+      configuration: new Configuration(),
+    })
+
+    assert.deepStrictEqual(checkWindowConfiguration.pageId, 'my-page')
+  })
+
   describe('toMatchSettings', () => {
     let mockDriver,
       driver,